Remove Safe interface
