+++ packages/safe-core-sdk/src/index.ts
@@ -2,8 +2,7 @@ import { ContractNetworksConfig } from './configuration/contracts'
 import EthAdapter from './ethereumLibs/EthAdapter'
 import EthersAdapter, { EthersAdapterConfig } from './ethereumLibs/EthersAdapter'
 import Web3Adapter, { Web3AdapterConfig } from './ethereumLibs/Web3Adapter'
-import EthersSafe from './EthersSafe'
-import Safe, { ConnectEthersSafeConfig, EthersSafeConfig } from './Safe'
+import EthersSafe, { ConnectEthersSafeConfig, EthersSafeConfig } from './EthersSafe'
 import { SafeSignature } from './utils/signatures/SafeSignature'
 import SafeTransaction, { SafeTransactionDataPartial } from './utils/transactions/SafeTransaction'
 
@@ -14,7 +13,6 @@ export {
   Web3Adapter,
   EthersAdapterConfig,
   EthersAdapter,
-  Safe,
   EthersSafeConfig,
   ConnectEthersSafeConfig,
   SafeSignature,
